# openapiint.yml
openapi: 3.0.3
info:
  title: API Interna de Logística - Configuración
  description: |
    API interna para la gestión de configuración del módulo de logística.
    Esta API es utilizada por el sistema interno para configurar tipos de transporte,
    zonas de cobertura y tarifas.
    
    **RF-001: Servicio de Configuración Base**
    - Gestión de tipos de transporte con velocidades promedio y tarifas base
    - Gestión de zonas de cobertura con códigos postales argentinos
    - Factor volumétrico configurable
  version: 1.0.0
  contact:
    name: Grupo 12 - Desarrollo de Software 2025
    email: soporte@logistica.com

servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
  - url: https://api-logistica.example.com
    description: Servidor de producción

tags:
  - name: Config - Transport Methods
    description: Gestión de métodos de transporte (RF-001)
  - name: Config - Coverage Zones
    description: Gestión de zonas de cobertura (RF-001)

paths:
  /config/transport-methods:
    get:
      tags:
        - Config - Transport Methods
      summary: Listar métodos de transporte
      description: Obtiene todos los métodos de transporte configurados con sus tarifas asociadas
      operationId: listTransportMethods
      responses:
        '200':
          description: Lista de métodos de transporte
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransportMethod'
              examples:
                success:
                  value:
                    - id: "ee4a3214-9563-4a7c-a0fc-cd39fdfac8a8"
                      code: "air"
                      name: "Aéreo"
                      description: "Transporte aéreo para envíos urgentes"
                      averageSpeed: 800
                      estimatedDays: "1-3"
                      baseCostPerKm: 0.8
                      baseCostPerKg: 5.0
                      isActive: true
                      createdAt: "2025-10-16T21:00:00Z"
                      updatedAt: "2025-10-16T21:00:00Z"
                      tariffConfigs:
                        - id: "24e545d1-a307-42fb-8d51-ddba014dd1ff"
                          baseTariff: 100.0
                          costPerKg: 8.0
                          costPerKm: 1.2
                          volumetricFactor: 200
                          environment: "development"
                          isActive: true

    post:
      tags:
        - Config - Transport Methods
      summary: Crear método de transporte
      description: Crea un nuevo método de transporte con sus configuraciones base
      operationId: createTransportMethod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransportMethodDto'
            examples:
              air:
                value:
                  code: "air"
                  name: "Aéreo"
                  description: "Transporte aéreo para envíos urgentes"
                  averageSpeed: 800
                  estimatedDays: "1-3"
                  baseCostPerKm: 0.8
                  baseCostPerKg: 5.0
                  isActive: true
      responses:
        '201':
          description: Método de transporte creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransportMethod'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Ya existe un método con ese código
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /config/transport-methods/{id}:
    get:
      tags:
        - Config - Transport Methods
      summary: Obtener método de transporte por ID
      description: Obtiene los detalles de un método de transporte específico
      operationId: getTransportMethod
      parameters:
        - name: id
          in: path
          required: true
          description: ID del método de transporte
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Método de transporte encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransportMethod'
        '404':
          description: Método de transporte no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Config - Transport Methods
      summary: Actualizar método de transporte
      description: Actualiza los datos de un método de transporte existente
      operationId: updateTransportMethod
      parameters:
        - name: id
          in: path
          required: true
          description: ID del método de transporte
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransportMethodDto'
      responses:
        '200':
          description: Método de transporte actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransportMethod'
        '404':
          description: Método de transporte no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflicto con código existente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /config/coverage-zones:
    get:
      tags:
        - Config - Coverage Zones
      summary: Listar zonas de cobertura
      description: Obtiene todas las zonas de cobertura configuradas
      operationId: listCoverageZones
      responses:
        '200':
          description: Lista de zonas de cobertura
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CoverageZone'
              examples:
                success:
                  value:
                    - id: "02ac9ed3-bf04-4b1c-8e92-1d3276bb86e5"
                      name: "Buenos Aires Capital"
                      description: "Capital Federal y zonas aledañas"
                      postalCodes: ["C1000", "C1001", "C1002", "C1003", "C1004", "C1005"]
                      isActive: true
                      createdAt: "2025-10-16T21:00:00Z"
                      updatedAt: "2025-10-16T21:00:00Z"

    post:
      tags:
        - Config - Coverage Zones
      summary: Crear zona de cobertura
      description: Crea una nueva zona de cobertura con códigos postales argentinos
      operationId: createCoverageZone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCoverageZoneDto'
            examples:
              buenosAires:
                value:
                  name: "Buenos Aires Capital"
                  description: "Capital Federal y zonas aledañas"
                  postalCodes: ["C1000", "C1001", "C1002"]
                  isActive: true
      responses:
        '201':
          description: Zona de cobertura creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoverageZone'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /config/coverage-zones/{id}:
    get:
      tags:
        - Config - Coverage Zones
      summary: Obtener zona de cobertura por ID
      description: Obtiene los detalles de una zona de cobertura específica
      operationId: getCoverageZone
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la zona de cobertura
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Zona de cobertura encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoverageZone'
        '404':
          description: Zona de cobertura no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Config - Coverage Zones
      summary: Actualizar zona de cobertura
      description: Actualiza los datos de una zona de cobertura existente
      operationId: updateCoverageZone
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la zona de cobertura
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCoverageZoneDto'
      responses:
        '200':
          description: Zona de cobertura actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoverageZone'
        '404':
          description: Zona de cobertura no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    TransportMethod:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID único del método de transporte
        code:
          type: string
          enum: [air, sea, rail, road]
          description: Código único del tipo de transporte
        name:
          type: string
          description: Nombre del método de transporte
        description:
          type: string
          nullable: true
          description: Descripción del método de transporte
        averageSpeed:
          type: integer
          description: Velocidad promedio en km/h
        estimatedDays:
          type: string
          description: Rango de días estimados de entrega (ej 1-3, 3-7)
        baseCostPerKm:
          type: number
          format: decimal
          description: Tarifa base por kilómetro
        baseCostPerKg:
          type: number
          format: decimal
          description: Tarifa base por kilogramo
        isActive:
          type: boolean
          description: Indica si el método está activo
        createdAt:
          type: string
          format: date-time
          description: Fecha de creación
        updatedAt:
          type: string
          format: date-time
          description: Fecha de última actualización
        tariffConfigs:
          type: array
          items:
            $ref: '#/components/schemas/TariffConfig'

    CreateTransportMethodDto:
      type: object
      required:
        - code
        - name
        - averageSpeed
        - estimatedDays
        - baseCostPerKm
        - baseCostPerKg
      properties:
        code:
          type: string
          enum: [air, sea, rail, road]
        name:
          type: string
          minLength: 2
          maxLength: 100
        description:
          type: string
        averageSpeed:
          type: integer
          minimum: 1
        estimatedDays:
          type: string
          maxLength: 20
        baseCostPerKm:
          type: number
          minimum: 0
        baseCostPerKg:
          type: number
          minimum: 0
        isActive:
          type: boolean
          default: true

    UpdateTransportMethodDto:
      type: object
      properties:
        code:
          type: string
          enum: [air, sea, rail, road]
        name:
          type: string
          minLength: 2
          maxLength: 100
        description:
          type: string
        averageSpeed:
          type: integer
          minimum: 1
        estimatedDays:
          type: string
          maxLength: 20
        baseCostPerKm:
          type: number
          minimum: 0
        baseCostPerKg:
          type: number
          minimum: 0
        isActive:
          type: boolean

    CoverageZone:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID único de la zona de cobertura
        name:
          type: string
          description: Nombre de la zona de cobertura
        description:
          type: string
          nullable: true
          description: Descripción de la zona de cobertura
        postalCodes:
          type: array
          items:
            type: string
          description: Array de códigos postales argentinos cubiertos
        isActive:
          type: boolean
          description: Indica si la zona está activa
        createdAt:
          type: string
          format: date-time
          description: Fecha de creación
        updatedAt:
          type: string
          format: date-time
          description: Fecha de última actualización

    CreateCoverageZoneDto:
      type: object
      required:
        - name
        - postalCodes
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        description:
          type: string
        postalCodes:
          type: array
          items:
            type: string
          minItems: 1
        isActive:
          type: boolean
          default: true

    UpdateCoverageZoneDto:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        description:
          type: string
        postalCodes:
          type: array
          items:
            type: string
          minItems: 1
        isActive:
          type: boolean

    TariffConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        transportMethodId:
          type: string
          format: uuid
        baseTariff:
          type: number
          format: decimal
          description: Tarifa base del servicio
        costPerKg:
          type: number
          format: decimal
          description: Costo por kilogramo
        costPerKm:
          type: number
          format: decimal
          description: Costo por kilómetro
        volumetricFactor:
          type: integer
          description: Factor volumétrico configurable para cálculo de peso volumétrico
        environment:
          type: string
          enum: [development, testing, production]
          description: Ambiente al que aplica esta configuración
        isActive:
          type: boolean
        validFrom:
          type: string
          format: date-time
        validTo:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string